requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
requests==2.32.3
